+++ ethergo/submitter/suite_test.go
@@ -119,14 +119,6 @@ func (s *SubmitterSuite) SetupSuite() {
 		}(i, chainID)
 	}
 	wg.Wait()
-
-	// fallback is currently untested. For now we disable it in tests.
-	// TODO: this should be fixed, or ideally, fallback can be removed.
-	og := submitter.ForceNoFallbackIfZero
-	submitter.SetForceNoFallback(true)
-	s.T().Cleanup(func() {
-		submitter.SetForceNoFallback(og)
-	})
 }
 
 // SetupTest sets up the signer and funds the account with 10 eth on each backend.
@@ -209,12 +201,6 @@ func (t *TXSubmitterDBSuite) SetupSuite() {
 	var err error
 	t.metrics, err = metrics.NewByType(t.GetSuiteContext(), buildInfo, metricsHandler)
 	t.Require().NoError(err)
-
-	og := submitter.ForceNoFallbackIfZero
-	submitter.SetForceNoFallback(true)
-	t.T().Cleanup(func() {
-		submitter.SetForceNoFallback(og)
-	})
 }
 
 func (t *TXSubmitterDBSuite) SetupTest() {