+++ packages/sdk-router/src/router/index.ts
@@ -1,5 +1,3 @@
-export * from './query'
-
 export * from './router'
 export * from './routerSet'